# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DueReviewListResponse": {
        "properties": {
          "course_id": {
            "title": "Course Id",
            "type": "integer"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/ReviewScheduleItemResponse"
            },
            "title": "Reviews",
            "type": "array"
          }
        },
        "required": [
          "course_id",
          "reviews"
        ],
        "title": "DueReviewListResponse",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "params": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonPrimitive"
            },
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "code"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonPrimitive": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "ReviewCompletionRequest": {
        "properties": {
          "course_id": {
            "exclusiveMinimum": 0.0,
            "title": "Course Id",
            "type": "integer"
          },
          "score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "number"
          },
          "topic": {
            "minLength": 1,
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "course_id",
          "topic",
          "score"
        ],
        "title": "ReviewCompletionRequest",
        "type": "object"
      },
      "ReviewScheduleItemResponse": {
        "properties": {
          "course_id": {
            "title": "Course Id",
            "type": "integer"
          },
          "difficulty": {
            "title": "Difficulty",
            "type": "number"
          },
          "interval_days": {
            "title": "Interval Days",
            "type": "integer"
          },
          "interval_index": {
            "title": "Interval Index",
            "type": "integer"
          },
          "is_due": {
            "title": "Is Due",
            "type": "boolean"
          },
          "last_reviewed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Reviewed At"
          },
          "next_review_at": {
            "title": "Next Review At",
            "type": "string"
          },
          "review_count": {
            "title": "Review Count",
            "type": "integer"
          },
          "score_at_last_review": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score At Last Review"
          },
          "stability": {
            "title": "Stability",
            "type": "number"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "topic",
          "course_id",
          "interval_index",
          "interval_days",
          "last_reviewed_at",
          "next_review_at",
          "score_at_last_review",
          "difficulty",
          "stability",
          "review_count",
          "is_due"
        ],
        "title": "ReviewScheduleItemResponse",
        "type": "object"
      },
      "ReviewScheduleListResponse": {
        "properties": {
          "course_id": {
            "title": "Course Id",
            "type": "integer"
          },
          "schedules": {
            "items": {
              "$ref": "#/components/schemas/ReviewScheduleItemResponse"
            },
            "title": "Schedules",
            "type": "array"
          }
        },
        "required": [
          "course_id",
          "schedules"
        ],
        "title": "ReviewScheduleListResponse",
        "type": "object"
      },
      "ReviewScheduleRequest": {
        "properties": {
          "course_id": {
            "exclusiveMinimum": 0.0,
            "title": "Course Id",
            "type": "integer"
          },
          "topic": {
            "minLength": 1,
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "course_id",
          "topic"
        ],
        "title": "ReviewScheduleRequest",
        "type": "object"
      },
      "ReviewScheduleResponse": {
        "properties": {
          "schedule": {
            "$ref": "#/components/schemas/ReviewScheduleItemResponse"
          }
        },
        "required": [
          "schedule"
        ],
        "title": "ReviewScheduleResponse",
        "type": "object"
      },
...
      "UpcomingReviewListResponse": {
        "properties": {
          "course_id": {
            "title": "Course Id",
            "type": "integer"
          },
          "days_ahead": {
            "title": "Days Ahead",
            "type": "integer"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/ReviewScheduleItemResponse"
            },
            "title": "Reviews",
            "type": "array"
          }
        },
        "required": [
          "course_id",
          "days_ahead",
          "reviews"
        ],
        "title": "UpcomingReviewListResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Sophia API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/review/complete": {
      "post": {
        "operationId": "completeReview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewScheduleResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Complete Review Schedule",
        "tags": [
          "review"
        ]
      }
    },
    "/api/review/due": {
      "get": {
        "operationId": "listDueReviews",
        "parameters": [
          {
            "in": "query",
            "name": "course_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Course Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DueReviewListResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "List Due Reviews",
        "tags": [
          "review"
        ]
      }
    },
    "/api/review/schedules": {
      "get": {
        "operationId": "listReviewSchedules",
        "parameters": [
          {
            "in": "query",
            "name": "course_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Course Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "topic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewScheduleListResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "List Review Schedules",
        "tags": [
          "review"
        ]
      },
      "post": {
        "operationId": "scheduleReview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewScheduleResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Create Review Schedule",
        "tags": [
          "review"
        ]
      }
    },
    "/api/review/upcoming": {
      "get": {
        "operationId": "listUpcomingReviews",
        "parameters": [
          {
            "in": "query",
            "name": "course_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Course Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "days_ahead",
            "required": false,
            "schema": {
              "default": 3,
              "maximum": 365,
              "minimum": 1,
              "title": "Days Ahead",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpcomingReviewListResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "List Upcoming Reviews",
        "tags": [
          "review"
        ]
      }
    },
...
  }
}